Remove upper limit check in dealer win condition
The condition for the dealer's win in the game has been simplified. We've removed the unnecessary check that verifies if the dealer's total is less than or equal to Blackjack, as it was redundant because it is checked earlier. Henceforth, the dealer will win if their total is simply greater than the player's total.