Split generating the dice rolls from checking that a straight was formed.
This eliminates a lot of duplicate checks (check dice rolls with 1 item, add item, check entire list again). This also limits the number of re-set prints as the same re-set isn't print multiple times (once for dice rolls with 1 item, again with dice roll with 2 items, etc) and it eliminates the checks for numbers that can't be added yet (a roll list with 1 item cannot contain all 6 numbers, so a lot of unnecessary re-sets are printed).